revised per review, refactored test for pytest
