+++ WrightTools/collection/_collection.py
@@ -112,14 +112,9 @@ class Collection(Group):
 
         for name in self.item_names:
             item = self[name]
-            if isinstance(item, wt_data.Data):
-                item.convert(units, convert_variables=convert_variables, verbose=verbose)
-            elif isinstance(item, Collection):
-                Collection.convert(
-                    item, units, convert_variables=convert_variables, verbose=verbose
-                )
-            else:
-                pass
+            item.convert(units, convert_variables=convert_variables, verbose=verbose)
+            
+            
 
     def create_data(self, name="data", position=None, **kwargs):
         """Create a new child data.
